Add files via upload
In this project you will use real data from the Titanic Tragedy to take a closer look at the people who survived this disaster and also predict the likelihood of survival.

On April 15, 1912, the largest passenger liner ever made collided with an iceberg during her maiden voyage. When the Titanic sank it killed 1502 out of 2224 passengers and crew. This sensational tragedy shocked the international community and led to better safety regulations for ships. One of the reasons that the shipwreck resulted in such loss of life was that there were not enough lifeboats for the passengers and crew. Although there was some element of luck involved in surviving the sinking, some groups of people were more likely to survive than others.

The titanic.csv file contains data of the real Titanic passengers. Each row represents one person. The column heading variables have the following meanings:

·         survival: Survival (0 = no; 1 = yes)

·         class: Passenger class (1 = first; 2 = second; 3 = third)

·         name: Name

·         sex: Sex

·         age: Age

·         sibsp: Number of siblings/spouses aboard

·         parch: Number of parents/children aboard

·         ticket: Ticket number

·         fare: Passenger fare

·         cabin: Cabin

·         embarked: Port of embarkation (C = Cherbourg; Q = Queenstown; S = Southampton)

·         boat: Lifeboat (if survived)

·         body: Body number (if did not survive and body was recovered)